requirements.txt
streamlit
requests
greek-stemmer-pos
python-dateutil
pandas
altair